[FEAT] 중복 이메일 가입 시 통합되도록 수정

+++ src/main/java/com/server/capple/config/security/jwt/service/JwtServiceImpl.java
@@ -99,8 +99,8 @@ public class JwtServiceImpl implements JwtService {
         return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("tokenType", String.class);
     }
 
-    public Integer getMemberId(String token) {
-        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("memberId", Integer.class);
+    public Long getMemberId(String token) {
+        return Jwts.parser().verifyWith(secretKey).build().parseSignedClaims(token).getPayload().get("memberId", Long.class);
     }
 
     public String getRole(String token) { //TODO 인가 관련 추가 구현 필요
